+++ src/m365/purview/commands.ts
@@ -2,8 +2,14 @@ const prefix: string = 'purview';
 
 export default {
   AUDITLOG_LIST: `${prefix} auditlog list`,
+  RETENTIONEVENT_GET: `${prefix} retentionevent get`,
+  RETENTIONEVENT_LIST: `${prefix} retentionevent list`,
   RETENTIONEVENT_REMOVE: `${prefix} retentionevent remove`,
+  RETENTIONEVENTTYPE_ADD: `${prefix} retentioneventtype add`,
   RETENTIONEVENTTYPE_GET: `${prefix} retentioneventtype get`,
+  RETENTIONEVENTTYPE_LIST: `${prefix} retentioneventtype list`,
+  RETENTIONEVENTTYPE_REMOVE: `${prefix} retentioneventtype remove`,
+  RETENTIONEVENTTYPE_SET: `${prefix} retentioneventtype set`,
   RETENTIONLABEL_ADD: `${prefix} retentionlabel add`,
   RETENTIONLABEL_GET: `${prefix} retentionlabel get`,
   RETENTIONLABEL_LIST: `${prefix} retentionlabel list`,