+++ src/m365/purview/commands/retentionevent/retentionevent-get.ts
@@ -1,0 +1,83 @@
+import { Logger } from '../../../../cli/Logger';
+import GraphCommand from '../../../base/GraphCommand';
+import GlobalOptions from '../../../../GlobalOptions';
+import commands from '../../commands';
+import { accessToken } from '../../../../utils/accessToken';
+import auth from '../../../../Auth';
+import request, { CliRequestOptions } from '../../../../request';
+import { validation } from '../../../../utils/validation';
+
+interface CommandArgs {
+  options: Options;
+}
+
+interface Options extends GlobalOptions {
+  id: string;
+}
+
+class PurviewRetentionEventGetCommand extends GraphCommand {
+  public get name(): string {
+    return commands.RETENTIONEVENT_GET;
+  }
+
+  public get description(): string {
+    return 'Retrieve the specified retention event';
+  }
+
+  constructor() {
+    super();
+
+    this.#initOptions();
+    this.#initValidators();
+  }
+
+  #initOptions(): void {
+    this.options.unshift(
+      {
+        option: '-i, --id <id>'
+      }
+    );
+  }
+
+  #initValidators(): void {
+    this.validators.push(
+      async (args: CommandArgs) => {
+        if (!validation.isValidGuid(args.options.id)) {
+          return `'${args.options.id}' is not a valid GUID.`;
+        }
+
+        return true;
+      }
+    );
+  }
+
+  public async commandAction(logger: Logger, args: CommandArgs): Promise<void> {
+    if (this.verbose) {
+      logger.logToStderr(`Retrieving retention event with id ${args.options.id}`);
+    }
+
+    try {
+      const isAppOnlyAccessToken: boolean | undefined = accessToken.isAppOnlyAccessToken(auth.service.accessTokens[this.resource].accessToken);
+
+      if (isAppOnlyAccessToken) {
+        throw 'This command currently does not support app only permissions.';
+      }
+
+      const requestOptions: CliRequestOptions = {
+        url: `${this.resource}/beta/security/triggers/retentionEvents/${args.options.id}`,
+        headers: {
+          accept: 'application/json;odata.metadata=none'
+        },
+        responseType: 'json'
+      };
+
+      const res: any = await request.get<any>(requestOptions);
+      logger.log(res);
+    }
+    catch (err: any) {
+      this.handleRejectedODataJsonPromise(err);
+    }
+  }
+}
+
+module.exports = new PurviewRetentionEventGetCommand();