+++ src/m365/spo/commands/listitem/listitem-add.ts
@@ -13,7 +13,7 @@ interface CommandArgs {
   options: Options;
 }
 
-interface Options extends GlobalOptions {
+export interface Options extends GlobalOptions {
   webUrl: string;
   listId?: string;
   listTitle?: string;
@@ -282,7 +282,8 @@ class SpoListItemAddCommand extends SpoCommand {
       'folder',
       'debug',
       'verbose',
-      'output'
+      'output',
+      '_'
     ];
 
     Object.keys(options).forEach(key => {